@@ -28,7 +28,10 @@ public class JwtRequestFilter extends OncePerRequestFilter {
     private JwtTokenUtil jwtTokenUtil;
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
+    protected void doFilterInternal(
+            HttpServletRequest request,
+            HttpServletResponse response,
+            FilterChain chain)
             throws ServletException, IOException {
 
         final String requestTokenHeader = request.getHeader("Authorization");
@@ -60,15 +63,14 @@ public class JwtRequestFilter extends OncePerRequestFilter {
             // authentication
             try {
                 if (jwtTokenUtil.validateToken(jwtToken, userDetails)) {
-
-                    UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(
-                            userDetails, null, userDetails.getAuthorities());
-                    usernamePasswordAuthenticationToken
-                            .setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                    // After setting the Authentication in the context, we specify
-                    // that the current user is authenticated. So it passes the
-                    // Spring Security Configurations successfully.
-                    SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
+                    // UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(
+                    //         userDetails, null, userDetails.getAuthorities());
+                    // usernamePasswordAuthenticationToken
+                    //         .setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+                    // // After setting the Authentication in the context, we specify
+                    // // that the current user is authenticated. So it passes the
+                    // // Spring Security Configurations successfully.
+                    // SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
                 } else {
                     System.out.println("Invalid/expired token");
                 }
